sched: Change rq->nr_running to unsigned int

+++ kernel/sched/sched.h
@@ -160,7 +160,7 @@ struct cfs_bandwidth { };
 
 struct cfs_rq {
 	struct load_weight load;
-	unsigned long nr_running, h_nr_running;
+	unsigned int nr_running, h_nr_running;
 
 	u64 exec_clock;
 	u64 min_vruntime;
@@ -212,7 +212,7 @@ static inline int rt_bandwidth_enabled(void)
 
 struct rt_rq {
 	struct rt_prio_array active;
-	unsigned long rt_nr_running;
+	unsigned int rt_nr_running;
 #if defined CONFIG_SMP || defined CONFIG_RT_GROUP_SCHED
 	struct {
 		int curr; 
@@ -263,7 +263,7 @@ struct rq {
 	
 	raw_spinlock_t lock;
 
-	unsigned long nr_running;
+	unsigned int nr_running;
 	#define CPU_LOAD_IDX_MAX 5
 	unsigned long cpu_load[CPU_LOAD_IDX_MAX];
 	unsigned long last_load_update_tick;
